backend/requirements.txt
fastapi
google-genai
pydantic>=2
python-dotenv
uvicorn[standard]
